pyproject.toml
[project]
name = "the-marionette"
version = "2.0.0"
description = "LLM-driven bulk user comment generation tool"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "pydantic>=2.7.0",
    "python-docx>=1.1.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.9.0",
]


[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "prompts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
